+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The present file will list all changes made to the project; according to the
 
 #### Added
 
+- `CommonDBTM::pre_addToDB()` added.
+
 #### Changes
 
 #### Deprecated